requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
pydantic[email]
pydantic-settings
python-dotenv
psycopg2-binary